move infoheadermodel to network package
